+++ pennyway-client-iOS/pennyway-client-iOS/View/TabView/ProfileView/ProfileMainView/ProfileMainView.swift
@@ -14,7 +14,7 @@ struct ProfileMainView: View {
                     
                     Spacer().frame(height: 9 * DynamicSizeFactor.factor())
                     
-                    ProfileSettingListView(userProfileViewModel: UserProfileViewModel())
+                    ProfileSettingListView()
                 }
                 .background(Color("Gray01"))
             }
